weixin

+++ ColdDrinkWeixin/Common/UserInfo.cs
@@ -16,7 +16,7 @@ namespace ColdDrinkWeixin.Common
         public string province { get; set; }
         public string language { get; set; }
         public string headimgurl { get; set; }
-        public int subscribe_time { get; set; }
+        public long subscribe_time { get; set; }
         public string unionid { get; set; }
         public string remark { get; set; }
         public int groupid { get; set; }
